+++ Carbon/Carbon.psm1.Import.Obsolete.ps1
@@ -5988,7 +5988,7 @@ function Invoke-CPowerShell
   </startup>
 </configuration>
 "@ -f $Runtime | Out-File -FilePath $activationConfigPath -Encoding OEM
-        Set-CEnvironmentVariable -Name $comPlusAppConfigEnvVarName -Value $activationConfigDir -ForProcess
+        Set-CEnvironmentVariable -Name $comPlusAppConfigEnvVarName -Value $activationConfigDir -ForProcess -NoWarn
     }
 
     $params = @{ }
@@ -6102,11 +6102,11 @@ function Invoke-CPowerShell
         {
             if( $originalCOMAppConfigEnvVar )
             {
-                Set-CEnvironmentVariable -Name $comPlusAppConfigEnvVarName -Value $originalCOMAppConfigEnvVar -ForProcess
+                Set-CEnvironmentVariable -Name $comPlusAppConfigEnvVarName -Value $originalCOMAppConfigEnvVar -ForProcess -NoWarn
             }
             else
             {
-                Remove-CEnvironmentVariable -Name $comPlusAppConfigEnvVarName -ForProcess
+                Remove-CEnvironmentVariable -Name $comPlusAppConfigEnvVarName -ForProcess -NoWarn
             }
         }
     }
@@ -6810,6 +6810,140 @@ filter Protect-CString
     }
 }
 
+
+
+function Remove-CEnvironmentVariable
+{
+    <#
+    .SYNOPSIS
+    Removes an environment variable.
+
+    .DESCRIPTION
+    Uses the .NET [Environment class](http://msdn.microsoft.com/en-us/library/z8te35sa) to remove an environment
+    variable from the Process, User, or Computer scopes.
+
+    Changes to environment variables in the User and Machine scope are not picked up by running processes.  Any running
+    processes that use this environment variable should be restarted.
+
+    Normally, you have to restart your PowerShell session/process to no longer see the variable in the `env:` drive. Use
+    the `-Force` switch to also remove the variable from the `env:` drive. This functionality was added in Carbon 2.3.0.
+
+    Beginning with Carbon 2.3.0, you can set an environment variable for a specific user by specifying the `-ForUser`
+    switch and passing the user's credentials with the `-Credential` parameter. This runs a separate PowerShell process
+    as that user to remove the variable.
+
+    Beginning in Carbon 2.3.0, you can specify multiple scopes from which to remove an environment variable. In previous
+    versions, you could only remove from one scope.
+
+    .LINK
+    Set-CEnvironmentVariable
+
+    .LINK
+    http://msdn.microsoft.com/en-us/library/z8te35sa
+
+    .EXAMPLE
+    Remove-CEnvironmentVariable -Name 'MyEnvironmentVariable' -ForProcess
+
+    Removes the `MyEnvironmentVariable` from the process scope.
+
+    .EXAMPLE
+    Remove-CEnvironmentVariable -Name 'SomeUsersVariable' -ForUser -Credential $credential
+
+    Demonstrates that you can remove another user's user-level environment variable by passing its credentials to the
+    `Credential` parameter. This runs a separate PowerShell process as that user to remove the variable.
+    #>
+    [CmdletBinding(SupportsShouldProcess)]
+    param(
+        [Parameter(Mandatory)]
+        # The environment variable to remove.
+        [String] $Name,
+
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        # Removes the environment variable for the current computer.
+        [switch] $ForComputer,
+
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        [Parameter(Mandatory, ParameterSetName='ForSpecificUser')]
+        # Removes the environment variable for the current user.
+        [switch] $ForUser,
+
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        # Removes the environment variable for the current process.
+        [switch] $ForProcess,
+
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        # Remove the variable from the current PowerShell session's `env:` drive, too. Normally, you have to restart
+        # your session to no longer see the variable in the `env:` drive.
+        #
+        # This parameter was added in Carbon 2.3.0.
+        [switch] $Force,
+
+        [Parameter(Mandatory, ParameterSetName='ForSpecificUser')]
+        # Remove an environment variable for a specific user.
+        [pscredential] $Credential,
+
+        [switch] $NoWarn
+    )
+
+    Set-StrictMode -Version 'Latest'
+    Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
+
+    if (-not $NoWarn)
+    {
+        Write-CRefactoredCommandWarning -CommandName $MyInvocation.MyCommand.Name -ModuleName 'Carbon.Environment'
+    }
+
+    if ($PSCmdlet.ParameterSetName -eq 'ForSpecificUser')
+    {
+        $parameters = $PSBoundParameters
+        $parameters.Remove('Credential')
+        $job = Start-Job -ScriptBlock {
+            Import-Module -Name (Join-Path -Path $using:carbonRoot -ChildPath 'Carbon.psd1')
+            $VerbosePreference = $using:VerbosePreference
+            $ErrorActionPreference = $using:ErrorActionPreference
+            $DebugPreference = $using:DebugPreference
+            $WhatIfPreference = $using:WhatIfPreference
+            Remove-CEnvironmentVariable @using:parameters
+        } -Credential $Credential
+        $job | Wait-Job | Receive-Job
+        $job | Remove-Job -Force -ErrorAction Ignore
+        return
+    }
+
+    if (-not $ForProcess -and -not $ForUser -and -not $ForComputer)
+    {
+        $msg = 'Environment variable target not specified. You must supply one of the ForComputer, ForUser, or ' +
+               'ForProcess switches.'
+        Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+        return
+    }
+
+    Invoke-Command -ScriptBlock {
+            if( $ForComputer )
+            {
+                [EnvironmentVariableTarget]::Machine
+            }
+
+            if( $ForUser )
+            {
+                [EnvironmentVariableTarget]::User
+            }
+
+            if( $ForProcess )
+            {
+                [EnvironmentVariableTarget]::Process
+            }
+        } |
+        Where-Object { $PSCmdlet.ShouldProcess( "${_}-level environment variable ""${Name}""", "remove" ) } |
+        ForEach-Object {
+                $scope = $_
+                [Environment]::SetEnvironmentVariable( $Name, [NullString]::Value, $scope )
+                if ($Force -and $scope -ne [EnvironmentVariableTarget]::Process)
+                {
+                    [Environment]::SetEnvironmentVariable($Name, [NullString]::Value, 'Process')
+                }
+            }
+}
 
 function Remove-CRegistryKeyValue
 {
@@ -7807,6 +7941,130 @@ function Revoke-CPrivilege
 }
 
 
+
+
+function Set-CEnvironmentVariable
+{
+    <#
+    .SYNOPSIS
+    Creates or sets an environment variable.
+
+    .DESCRIPTION
+    Uses the .NET [Environment class](http://msdn.microsoft.com/en-us/library/z8te35sa) to create or set an environment variable in the Process, User, or Machine scopes.
+
+    Changes to environment variables in the User and Machine scope are not picked up by running processes.  Any running processes that use this environment variable should be restarted.
+
+    Beginning with Carbon 2.3.0, you can set an environment variable for a specific user by specifying the `-ForUser` switch and passing the user's credentials with the `-Credential` parameter. This will run a PowerShell process as that user in order to set the environment variable.
+
+    Normally, you have to restart your PowerShell session/process to see the variable in the `env:` drive. Use the `-Force` switch to also add the variable to the `env:` drive. This functionality was added in Carbon 2.3.0.
+
+    .LINK
+    Remove-CEnvironmentVariable
+
+    .LINK
+    http://msdn.microsoft.com/en-us/library/z8te35sa
+
+    .EXAMPLE
+    Set-CEnvironmentVariable -Name 'MyEnvironmentVariable' -Value 'Value1' -ForProcess
+
+    Creates the `MyEnvironmentVariable` with an initial value of `Value1` in the process scope, i.e. the variable is only accessible in the current process.
+
+    .EXAMPLE
+    Set-CEnvironmentVariable -Name 'MyEnvironmentVariable' -Value 'Value1' -ForComputer
+
+    Creates the `MyEnvironmentVariable` with an initial value of `Value1` in the machine scope, i.e. the variable is accessible in all newly launched processes.
+
+    .EXAMPLE
+    Set-CEnvironmentVariable -Name 'SomeUsersEnvironmentVariable' -Value 'SomeValue' -ForUser -Credential $userCreds
+
+    Demonstrates that you can set a user-level environment variable for another user by passing its credentials to the `Credential` parameter. Runs a separate PowerShell process as that user to set the environment variable.
+    #>
+    [CmdletBinding(SupportsShouldProcess=$true)]
+    param(
+        [Parameter(Mandatory=$true)]
+        # The name of environment variable to add/set.
+        [string]$Name,
+
+        [Parameter(Mandatory=$true)]
+        # The environment variable's value.
+        [string]$Value,
+
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        # Sets the environment variable for the current computer.
+        [Switch]$ForComputer,
+
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        [Parameter(Mandatory=$true,ParameterSetName='ForSpecificUser')]
+        # Sets the environment variable for the current user.
+        [Switch]$ForUser,
+
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        # Sets the environment variable for the current process.
+        [Switch]$ForProcess,
+
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        # Set the variable in the current PowerShell session's `env:` drive, too. Normally, you have to restart your session to see the variable in the `env:` drive.
+        #
+        # This parameter was added in Carbon 2.3.0.
+        [Switch]$Force,
+
+        [Parameter(Mandatory=$true,ParameterSetName='ForSpecificUser')]
+        # Set an environment variable for a specific user.
+        [pscredential]$Credential,
+
+        [switch] $NoWarn
+    )
+
+    Set-StrictMode -Version 'Latest'
+    Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
+
+    if (-not $NoWarn)
+    {
+        Write-CRefactoredCommandWarning -CommandName $MyInvocation.MyCommand.Name -ModuleName 'Carbon.Environment'
+    }
+
+    if( $PSCmdlet.ParameterSetName -eq 'ForSpecificUser' )
+    {
+        $parameters = $PSBoundParameters
+        $parameters.Remove('Credential')
+        $job = Start-Job -ScriptBlock {
+            Import-Module -Name (Join-Path -path $using:carbonRoot -ChildPath 'Carbon.psd1' -Resolve)
+            $VerbosePreference = $using:VerbosePreference
+            $ErrorActionPreference = $using:ErrorActionPreference
+            $DebugPreference = $using:DebugPreference
+            $WhatIfPreference = $using:WhatIfPreference
+            Set-CEnvironmentVariable @using:parameters
+        } -Credential $Credential
+        $job | Wait-Job | Receive-Job
+        $job | Remove-Job -Force -ErrorAction Ignore
+        return
+    }
+
+    if( -not $ForProcess -and -not $ForUser -and -not $ForComputer )
+    {
+        Write-Error -Message ('Environment variable target not specified. You must supply one of the ForComputer, ForUser, or ForProcess switches.')
+        return
+    }
+
+    Invoke-Command -ScriptBlock {
+            if( $ForComputer )
+            {
+                [EnvironmentVariableTarget]::Machine
+            }
+
+            if( $ForUser )
+            {
+                [EnvironmentVariableTarget]::User
+            }
+
+            if( $Force -or $ForProcess )
+            {
+                [EnvironmentVariableTarget]::Process
+            }
+        } |
+        Where-Object { $PSCmdlet.ShouldProcess( "$_-level environment variable '$Name'", "set") } |
+        ForEach-Object { [Environment]::SetEnvironmentVariable( $Name, $Value, $_ ) }
+}
 
 function Set-CRegistryKeyValue
 {